requirements.txt
streamlit>=1.30.0
openai>=1.10.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.7.4
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.24.0
sympy>=1.12
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
audio-recorder-streamlit>=0.0.8
